Added footer to Admin Panel
I re-enabled the footer disabled by Latuminggi's LittleLink Admin fork. This footer adds a Home button which brings you to the landing page, in addition to Terms, Privacy and Contact buttons.

I also edited the policy pages for the buttons named above to include the new logo as well as the dark mode detection.