Fixed 'Edit User' changes not applying
Fixed bug where if a user was edited on the Admin Panel under 'Admin>Users> Edit User' changes would not apply without every value being changed. 

This bug originates from LittleLink Admin (the fork of which this fork is based on).

One part of this issue was that most fields were required to fill in. If only one would be left, empty, none would apply. The problem was if this requirement was removed, the password would always overwrite the current password, even when left empty. 
This meant you could only edit users if you knew their password, otherwise their passwords would be changed, probably preventing that user from login in ever again.

After some experimenting, I implemented the easiest fix I could come up with. A simple if-else statement that checks if the password field is empty and if so only saves the other parameters without touching the password. If the password is changed, the changes will be applied normally. 

Also fixed the profile image size for the default placeholder image.