Added check, if added links begin with "https"
Added check added, or edited links begin with https. Previously if links were saved without this formatting applied an error exception would be thrown.

For this I added two functions, one can check if a string ends with a certain string, the other checks if it ends with a certain string. The first function is used for the newly added check, the other one is planned to be used in a similar check in the future.